Move rating filtering tests
